% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_condition_codes.R
\name{add_condition_type_codes}
\alias{add_condition_type_codes}
\title{Adds condition type codes to provided data frame according to the \code{CONDITION_CODE} column}
\usage{
add_condition_type_codes(data)
}
\description{
Adds condition type codes to provided data frame according to the \code{CONDITION_CODE} column
}
