% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fate_codes.R
\name{add_fate_type_codes}
\alias{add_fate_type_codes}
\title{Adds fate type codes to provided data frame according to the \code{FATE_CODE} column}
\usage{
add_fate_type_codes(data)
}
\description{
Adds fate type codes to provided data frame according to the \code{FATE_CODE} column
}
