% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fishery_codes.R
\name{add_fishery_codes}
\alias{add_fishery_codes}
\title{Adds standard fishery codes to provided data frame according to the \code{FISHERY_GROUP} and \code{CATCH_SCHOOL_TYPE_CODE} columns}
\usage{
add_fishery_codes(data)
}
\description{
Adds standard fishery codes to provided data frame according to the \code{FISHERY_GROUP} and \code{CATCH_SCHOOL_TYPE_CODE} columns
}
