% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fishery_codes.R
\name{add_fishery_group_codes}
\alias{add_fishery_group_codes}
\title{Adds standard fishery group codes to provided data frame according to the \code{FISHERY_GROUP} column}
\usage{
add_fishery_group_codes(data)
}
\description{
Adds standard fishery group codes to provided data frame according to the \code{FISHERY_GROUP} column
}
