% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fleet_codes.R
\name{add_fleet_names_}
\alias{add_fleet_names_}
\title{DEPRECATED
Adds the \code{FLEET} (fleet english name), \code{FLEET_SORT} (fleet sorting order), \code{CPC_CODE} (fleet CPC code) and \code{STATUS} (fleet CPC status) to a data table already containing a \code{FLEET_CODE} column}
\usage{
add_fleet_names_(data, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\description{
DEPRECATED
Adds the \code{FLEET} (fleet english name), \code{FLEET_SORT} (fleet sorting order), \code{CPC_CODE} (fleet CPC code) and \code{STATUS} (fleet CPC status) to a data table already containing a \code{FLEET_CODE} column
}
