% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_species_codes.R
\name{add_species_metadata}
\alias{add_species_metadata}
\title{Adds various species' metadata to a dataset containing a \code{SPECIES_CODE} column}
\usage{
add_species_metadata(data, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{data}{The dataset}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
The original dataset augmented with the species metadata if that's the case
}
\description{
Adds various species' metadata to a dataset containing a \code{SPECIES_CODE} column
}
