% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_helpers.R
\name{all_codes}
\alias{all_codes}
\title{Extracts (and caches) all details of a metadata codelist table as available on \code{IOTDB}}
\usage{
all_codes(codelist = "FLEETS", connection = DB_IOTDB())
}
\arguments{
\item{codelist}{The name of a codelist table}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the content of the codelist table as a data.table
}
\description{
Extracts (and caches) all details of a metadata codelist table as available on \code{IOTDB}
}
