% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fleet_codes.R
\name{all_fleets}
\alias{all_fleets}
\title{Extracts all fleet codes and names from the \code{[meta].FLEETS} table on \code{\link{IOTDB}}}
\usage{
all_fleets(connection = DB_IOTDB())
}
\arguments{
\item{connection}{A connection to the IOTDB}
}
\value{
a data frame with multiple rows (one for each fleet) and the following columns:
\code{SORT} (the fleet sorting order)
\code{CODE} (the fleet code, e.g., AUS, EUFRA)
\code{CPC_CODE} (the CPC code for the fleet, e.g., AUS, EU)
\code{FLEET_NAME_EN} (the fleet english name)
\code{STATUS} (the status of the fleet CPC: one among \code{CPC}, \code{NON_CPC}, \code{NEI})
}
\description{
Extracts all fleet codes and names from the \code{[meta].FLEETS} table on \code{\link{IOTDB}}
}
