% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_helpers.R
\name{all_meta}
\alias{all_meta}
\title{Extracts all details of a metadata table as available on \code{IOTDB}}
\usage{
all_meta(table = "SPECIES", connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{table}{The table}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the content of the table as a data.table
}
\description{
Extracts all details of a metadata table as available on \code{IOTDB}
}
