% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R,
%   R/iotc_base_common_data_mgmt_species_codes.R
\name{all_species}
\alias{all_species}
\title{Extracts all details of the \code{[meta].SPECIES} table as available on \code{IOTDB}}
\usage{
all_species(connection = iotc.core.db.connections::getDefaultDBIHandler()())

all_species(connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{connection}{A connection to the IOTDB}
}
\value{
a data frame with multiple rows (one for each species) and the required columns to model the structure of the table

a data frame with multiple rows (one for each species) and the required columns to model the structure of the table
}
\description{
Extracts all details of the \code{[meta].SPECIES} table as available on \code{IOTDB}

Extracts all details of the \code{[meta].SPECIES} table as available on \code{IOTDB}
}
