% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{all_species_categories}
\alias{all_species_categories}
\title{Extracts all details of the \code{[meta].SPECIES_CATEGORIES} table as available on \code{IOTDB}}
\usage{
all_species_categories(connection = DB_IOTDB())
}
\arguments{
\item{connection}{A connection to the \code{\link{IOTDB}}}
}
\value{
a data frame with multiple rows (one for each species category) and the required columns to model the structure of the table
}
\description{
Extracts all details of the \code{[meta].SPECIES_CATEGORIES} table as available on \code{IOTDB}
}
