% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_buoys.R
\name{BU_data}
\alias{BU_data}
\title{Retrieves monthly effort data from the \code{V_ROS_EF} view on \code{\link{ROS_ANALYSIS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria}
\usage{
BU_data(
  years = NULL,
  date_from = NULL,
  date_to = NULL,
  fleet_codes = NULL,
  vessel_names = NULL,
  buoy_brands = NULL,
  buoy_models = NULL,
  buoy_IDs = NULL,
  longitude_range = NULL,
  latitude_range = NULL,
  IO_area_code = NULL,
  intersection_level = NULL,
  CWP_1_grids = NULL,
  CWP_5_grids = NULL,
  connection = DB_BUOYS(),
  IOTCStatistics_connection = DB_IOTCSTATISTICS()
)
}
\arguments{
\item{years}{A vector of years the extraction should be limited to}

\item{date_from}{A starting date (as '\code{yyyy-MM-dd}') the extraction should be limited to}

\item{date_to}{An ending date (\code{yyyy-MM-dd}') the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{vessel_names}{A vector of vessel names the extraction should be limited to}

\item{buoy_brands}{A vector of buoy brands the extraction should be limited to}

\item{buoy_models}{A vector of buoy models the extraction should be limited to}

\item{buoy_IDs}{A vector of buoy IDs the extraction should be limited to}

\item{longitude_range}{A vector of min-max longitudes (as decimal degrees) the extraction should be limited to}

\item{latitude_range}{A vector of min-max latitudes (as decimal degrees) the extraction should be limited to}

\item{IO_area_code}{An IOTC area code the extraction should be limited to. It can be a CWP grid code or an irregular grid code. The intersection is approximated using the 1x1 grid grids overlaping with the given area}

\item{intersection_level}{When an IOTC area code is set, limits the results to those points that are in grids whose intersection with the area amounts to the provided level (in the 0 to 1 range)}

\item{CWP_1_grids}{A vector of CWP 1x1 grid codes the the extraction should be limited to.}

\item{CWP_5_grids}{A vector of CWP 5x5 grid codes the the extraction should be limited to.}

\item{connection}{An ODBC connection to the \code{\link{BUOY_DATA}} database}
}
\value{
a data frame with the records from \code{BUOY_DATA} matching the filtering criteria
}
\description{
Retrieves monthly effort data from the \code{V_ROS_EF} view on \code{\link{ROS_ANALYSIS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria
}
\examples{
BU_DATA(2020, fleet_codes = "EUESP", longitude_range = c(40, 50), latitude_range = c(-15, -25))
BU_DATA(2020, fleet_codes = "EUESP", IO_area_code = "IRSYCEZ")
}
