% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_CATCH_UNIT_CODE}
\alias{C_CATCH_UNIT_CODE}
\title{The catch unit code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_CATCH_UNIT_CODE
}
\description{
The catch unit code column constant
}
\keyword{datasets}
