% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_CLASS_HIGH}
\alias{C_CLASS_HIGH}
\title{The class high column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_CLASS_HIGH
}
\description{
The class high column constant
}
\keyword{datasets}
