% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_CONDITION}
\alias{C_CONDITION}
\title{The condition column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_CONDITION
}
\description{
The condition column constant
}
\keyword{datasets}
