% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_END_LAT}
\alias{C_END_LAT}
\title{The end lat column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_END_LAT
}
\description{
The end lat column constant
}
\keyword{datasets}
