% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FATE}
\alias{C_FATE}
\title{The fate column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FATE
}
\description{
The fate column constant
}
\keyword{datasets}
