% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FATE_CODE}
\alias{C_FATE_CODE}
\title{The fate code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FATE_CODE
}
\description{
The fate code column constant
}
\keyword{datasets}
