% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FISHERY}
\alias{C_FISHERY}
\title{The fishery column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FISHERY
}
\description{
The fishery column constant
}
\keyword{datasets}
