% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FISHERY_GROUP_CODE}
\alias{C_FISHERY_GROUP_CODE}
\title{The fishery group code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FISHERY_GROUP_CODE
}
\description{
The fishery group code column constant
}
\keyword{datasets}
