% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FISHERY_TYPE}
\alias{C_FISHERY_TYPE}
\title{The fishery type column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FISHERY_TYPE
}
\description{
The fishery type column constant
}
\keyword{datasets}
