% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FLEET}
\alias{C_FLEET}
\title{The fleet column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FLEET
}
\description{
The fleet column constant
}
\keyword{datasets}
