% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_FLEET_CODE}
\alias{C_FLEET_CODE}
\title{The fleet code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_FLEET_CODE
}
\description{
The fleet code column constant
}
\keyword{datasets}
