% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_GEAR}
\alias{C_GEAR}
\title{The gear column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_GEAR
}
\description{
The gear column constant
}
\keyword{datasets}
