% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_IS_SPECIES_AGGREGATE}
\alias{C_IS_SPECIES_AGGREGATE}
\title{The 'is species aggregate' column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_IS_SPECIES_AGGREGATE
}
\description{
The 'is species aggregate' column constant
}
\keyword{datasets}
