% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_LENGTH_MEASURE_TYPE_CODE}
\alias{C_LENGTH_MEASURE_TYPE_CODE}
\title{The length measure type code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_LENGTH_MEASURE_TYPE_CODE
}
\description{
The length measure type code column constant
}
\keyword{datasets}
