% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_MONTH_START}
\alias{C_MONTH_START}
\title{The month (start) column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_MONTH_START
}
\description{
The month (start) column constant
}
\keyword{datasets}
