% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_NUM_INTERACTIONS}
\alias{C_NUM_INTERACTIONS}
\title{The num. interactions column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_NUM_INTERACTIONS
}
\description{
The num. interactions column constant
}
\keyword{datasets}
