% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_QUALITY}
\alias{C_QUALITY}
\title{The 'Quality' column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_QUALITY
}
\description{
The 'Quality' column constant
}
\keyword{datasets}
