% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_RAISING}
\alias{C_RAISING}
\title{The raising column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_RAISING
}
\description{
The raising column constant
}
\keyword{datasets}
