% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_SPECIES_CATEGORY}
\alias{C_SPECIES_CATEGORY}
\title{The species category column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_SPECIES_CATEGORY
}
\description{
The species category column constant
}
\keyword{datasets}
