% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_SPECIES_CODE}
\alias{C_SPECIES_CODE}
\title{The species code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_SPECIES_CODE
}
\description{
The species code column constant
}
\keyword{datasets}
