% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_SPECIES_GROUP}
\alias{C_SPECIES_GROUP}
\title{The species group column constant}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
C_SPECIES_GROUP

C_SPECIES_GROUP
}
\description{
The species group column constant

The species group column constant
}
\keyword{datasets}
