% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_START_LON}
\alias{C_START_LON}
\title{The start lon column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_START_LON
}
\description{
The start lon column constant
}
\keyword{datasets}
