% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_START_TIME}
\alias{C_START_TIME}
\title{The start time column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_START_TIME
}
\description{
The start time column constant
}
\keyword{datasets}
