% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{C_TOT_NUM_FISH}
\alias{C_TOT_NUM_FISH}
\title{The tot num fish column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
C_TOT_NUM_FISH
}
\description{
The tot num fish column constant
}
\keyword{datasets}
