% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{CA_raw_std}
\alias{CA_raw_std}
\title{Same as \code{CA_raw(use_official = FALSE, ...)}}
\usage{
CA_raw_std(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  catch_unit_codes = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_result = TRUE
)
}
\description{
Same as \code{CA_raw(use_official = FALSE, ...)}
}
