% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_ROS.R
\name{CA_ros}
\alias{CA_ros}
\title{Retrieves monthly catch data from the  \code{V_ROS_CA} view on \code{\link{ROS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria}
\usage{
CA_ros(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  catch_unit_codes = NULL,
  fate_codes = NULL,
  connection = DB_ROS(),
  IOTDB_connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\arguments{
\item{years}{A vector of years the extraction should be limited to}

\item{flag_codes}{A vector of flag codes the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to#'}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{fishery_group_codes}{A vector of fishery group codes the extraction should be limited to}

\item{school_type_codes}{A vector of school type codes the extraction should be limited to}

\item{species_codes}{A vector of species codes the extraction should be limited to}

\item{species_category_codes}{A vector of species category codes (BILLFISH, SEERFISH, NERITIC, TEMPERATE, TROPICAL, TUNAS_NEI, SHARKS, MANTAS, SEABIRDS, CETACEANS, TURTLES, OTHERS) the extraction should be limited to}

\item{species_group_codes}{A vector of species group codes (BILLFISH, SEERFISH, TUNAS, SHARKS, OTHERS) the extraction should be limited to}

\item{species_wp_codes}{A vector of species WP codes (BILL, NERI, TEMP, TROP, BYCT, UNCL) the extraction should be limited to}

\item{catch_unit_codes}{A vector of catch unit codes the extraction should be limited to (currently, \code{NO} and \code{MT}, as the conversion from KG is embedded)}

\item{fate_codes}{A vector of fate codes the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{ROS}} database}

\item{IOTDB_connection}{An ODBC connection to the \code{\link{IOTDB}} database}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}
}
\value{
a data frame with the records from \code{V_ROS_CA} matching the filtering criteria
}
\description{
Retrieves monthly catch data from the  \code{V_ROS_CA} view on \code{\link{ROS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria
}
\examples{
CA_ros(years = seq(1980, 2000, 1), fishery_group_codes = "PS", catch_unit_codes = c("MT"))
}
