% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{CATCH}
\alias{CATCH}
\title{The catch column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATCH
}
\description{
The catch column constant
}
\keyword{datasets}
