% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{CATCH_IN_NUMBERS}
\alias{CATCH_IN_NUMBERS}
\title{The catch in numbers column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATCH_IN_NUMBERS
}
\description{
The catch in numbers column constant
}
\keyword{datasets}
