% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{CATCH_SCHOOL_TYPE_CODE}
\alias{CATCH_SCHOOL_TYPE_CODE}
\title{The catch school type code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
CATCH_SCHOOL_TYPE_CODE
}
\description{
The catch school type code column constant
}
\keyword{datasets}
