% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{CE_raw_off}
\alias{CE_raw_off}
\title{Same as \code{CE_raw(use_official = TRUE, ...)}}
\usage{
CE_raw_off(
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  effort_unit_codes = NULL,
  catch_unit_codes = NULL,
  connection = DB_IOTDB(),
  factorize_result = TRUE
)
}
\description{
Same as \code{CE_raw(use_official = TRUE, ...)}
}
