% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{CLASS_LOW}
\alias{CLASS_LOW}
\title{The class low column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
CLASS_LOW
}
\description{
The class low column constant
}
\keyword{datasets}
