% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_lw_equations.R
\name{convert_measure}
\alias{convert_measure}
\title{Converts a size measurement to another for a given species (if a corresponding equation is available)}
\usage{
convert_measure(
  species_code,
  source_measure_code,
  target_measure_code,
  source_measure,
  fishery_group_code = NA,
  connection = iotc.core.db.connections::getDefaultDBIHandler()()
)
}
\arguments{
\item{species_code}{The species code}

\item{source_measure_code}{The code identifying the type of source size measurement}

\item{target_measure_code}{The code identifying the type of target size measurement}

\item{source_measure}{The source measurement to convert}

\item{fishery_group_code}{A fishery group code, in case specific convertion equations for the species exist}
}
\value{
the source measurement converted to the target size measurement
}
\description{
Converts a size measurement to another for a given species (if a corresponding equation is available)
}
\examples{
convert_measure("BET", "FL", "KG", 100) #Converts a fork length of 100 cm to rounded weight for a BET
convert_measure("BET", "FL", "KG", 100, "PS") #Converts a fork length of 100 cm to rounded weight for a BET according to PS-specific equations
}
