% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_quality.R
\name{data.quality}
\alias{data.quality}
\title{Same as \code{\link{data_quality}}}
\usage{
data.quality(
  year_from = NULL,
  year_to = NULL,
  fleet_code = NULL,
  gear_code = NULL,
  species_code = NULL,
  species_group_code = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_result = TRUE
)
}
\description{
Same as \code{\link{data_quality}}
}
