% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_quality.R
\name{data_quality}
\alias{data_quality}
\title{Produces a data-quality report (by dataset) for a given species / species group among the IOTC ones, optionally
filtered by years, fleet and fishery}
\usage{
data_quality(
  year_from = NULL,
  year_to = NULL,
  fleet_code = NULL,
  gear_code = NULL,
  species_code = NULL,
  species_group_code = NULL,
  connection = DB_IOTDB(),
  factorize_result = TRUE
)
}
\arguments{
\item{year_from}{The ending year}

\item{fleet_code}{A fleet code}

\item{gear_code}{A gear code}

\item{species_code}{A species code}

\item{species_group_code}{A species group code (among \code{BILL}, \code{NERI}, \code{TEMP}, \code{TROP})}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}
}
\value{
a data table with the quality reports for the strata selected through the filtering criteria above
}
\description{
Produces a data-quality report (by dataset) for a given species / species group among the IOTC ones, optionally
filtered by years, fleet and fishery
}
\examples{
data_quality(year_from = 1969, year_to = 2019, species_group_code = "TROP")
}
