% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_helpers.R
\name{decorate}
\alias{decorate}
\title{Due to a bug in my current version of data.table, you need to ensure the following
patch is applied:}
\usage{
decorate(
  data,
  factorize = TRUE,
  remove_non_standard_columns = TRUE,
  connection = DB_IOTDB()
)
}
\arguments{
\item{data}{A dataset to decorate}

\item{factorize}{If \code{TRUE}, factorization will be applied}

\item{remove_non_standard_columns}{If \code{TRUE}, non-standard columns (i.e., not belonging to the data model) will be removed.}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\description{
remotes::install_github("Rdatatable/data.table#4803")
}
\details{
otherwise, after the factorization, accessing the data table in some circumstances
yields the following error:

Error in rbindlist(list(head(x, topn), tail(x, topn)), use.names = FALSE) : STRING_PTR() can only be applied to a 'character', not a 'NULL'
}
