% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{EF_raw}
\alias{EF_raw}
\title{Retrieves *raw* monthly effort data from the 'standard' \code{V_LEGACY_EF_STD} or 'official' \code{V_LEGACY_EF_OFF} view on IOTDB, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria}
\usage{
EF_raw(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  effort_unit_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\arguments{
\item{use_official}{A flag that tells whether using the official (if TRUE) or standard (if FALSE) view}

\item{years}{A vector of years the extraction should be limited to}

\item{flag_codes}{A vector of flag codes the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to#'}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{fishery_group_codes}{A vector of fishery group codes the extraction should be limited to}

\item{school_type_codes}{A vector of school type codes the extraction should be limited to}

\item{effort_unit_codes}{A vector of effort unit codes the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{IOTDB}} database}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}
}
\value{
a data frame with the records from \code{V_LEGACY_EF} matching the filtering criteria
}
\description{
The 'standard' version includes the 'FLAG_CODE' column as a combination of the codes from both
the flag country and the reporting country for the record, and the 'FLEET_CODE' column as it derives
from the combination of FLAG_CODE and fishing gear mapped by the \code{CountryStratVsFleet}
(which turns several flag + reporting country into generic fleets such as \code{NEICE}, \code{NEIFR}, \code{NEIPS} and similar).
The 'official' version only includes the 'FLEET_CODE' column determined as indicated above.
}
\examples{
EF_raw(years = seq(1980, 2000, 1), fishery_group_codes = "PS", effort_unit_codes = c("FHOURS", "FDAYS"))
EF_raw(use_official = FALSE, years = seq(1980, 2000, 1), fishery_group_codes = "PS", effort_unit_codes = c("FHOURS", "FDAYS"))
}
