% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{EF_raw_off}
\alias{EF_raw_off}
\title{Same as \code{EF_raw(use_official = TRUE, ...)}}
\usage{
EF_raw_off(
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  effort_unit_codes = NULL,
  connection = DB_IOTDB(),
  factorize_result = TRUE
)
}
\description{
Same as \code{EF_raw(use_official = TRUE, ...)}
}
