% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{EF_raw_std}
\alias{EF_raw_std}
\title{Same as \code{EF_std(use_official = FALSE, ...)}}
\usage{
EF_raw_std(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  effort_unit_codes = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_result = TRUE
)
}
\description{
Same as \code{EF_std(use_official = FALSE, ...)}
}
