% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_ROS.R
\name{EF_ros}
\alias{EF_ros}
\title{Retrieves monthly effort data from the \code{V_ROS_EF} view on \code{\link{ROS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria}
\usage{
EF_ros(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  effort_unit_codes = NULL,
  connection = DB_ROS(),
  IOTDB_connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\arguments{
\item{years}{A vector of years the extraction should be limited to}

\item{flag_codes}{A vector of flag codes the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to#'}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{fishery_group_codes}{A vector of fishery group codes the extraction should be limited to}

\item{school_type_codes}{A vector of school type codes the extraction should be limited to}

\item{effort_unit_codes}{A vector of effort unit codes the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{ROS}} database}

\item{IOTDB_connection}{An ODBC connection to the \code{\link{IOTDB}} database}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}
}
\value{
a data frame with the records from \code{V_ROS_EF} matching the filtering criteria
}
\description{
Retrieves monthly effort data from the \code{V_ROS_EF} view on \code{\link{ROS}}, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria
}
\examples{
EF_ros(years = seq(1980, 2000, 1), fishery_group_codes = "PS", effort_unit_codes = c("SET"))
}
