% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{EFFORT}
\alias{EFFORT}
\title{The effort column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
EFFORT
}
\description{
The effort column constant
}
\keyword{datasets}
