% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{EFFORT_SCHOOL_TYPE_CODE}
\alias{EFFORT_SCHOOL_TYPE_CODE}
\title{The effort school type code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
EFFORT_SCHOOL_TYPE_CODE
}
\description{
The effort school type code column constant
}
\keyword{datasets}
