% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{EFFORT_UNIT_CODE}
\alias{EFFORT_UNIT_CODE}
\title{The effort unit code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
EFFORT_UNIT_CODE
}
\description{
The effort unit code column constant
}
\keyword{datasets}
