% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{END_LON}
\alias{END_LON}
\title{The end lon column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
END_LON
}
\description{
The end lon column constant
}
\keyword{datasets}
