% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{END_TIME}
\alias{END_TIME}
\title{The end time column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
END_TIME
}
\description{
The end time column constant
}
\keyword{datasets}
