% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_est.R
\name{est.SF}
\alias{est.SF}
\title{Alias for \code{\link{SF_est}}}
\usage{
est.SF(
  species_code,
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{SF_est}}
}
