% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_condition_types}
\alias{factorize_condition_types}
\title{Factorizes the condition type codes in a data frame}
\usage{
factorize_condition_types(to_factorize, connection = DB_IOTDB())
}
\arguments{
\item{to_factorize}{The data frame whose \code{CONDITION_TYPE_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{CONDITION_TYPE_CODE} factorized according to the default sequence of condition type codes
}
\description{
Factorizes the condition type codes in a data frame
}
