% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_conditions}
\alias{factorize_conditions}
\title{Factorizes the condition codes in a data frame}
\usage{
factorize_conditions(to_factorize, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{to_factorize}{The data frame whose \code{CONDITION_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{CONDITION_CODE} factorized according to the default sequence of condition codes
}
\description{
Factorizes the condition codes in a data frame
}
