% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fate_types}
\alias{factorize_fate_types}
\title{Factorizes the fate type codes in a data frame}
\usage{
factorize_fate_types(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{FATE_TYPE_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{FATE_TYPE_CODE} factorized according to the default sequence of fate type codes
}
\description{
Factorizes the fate type codes in a data frame
}
