% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fates}
\alias{factorize_fates}
\title{Factorizes the fate codes in a data frame}
\usage{
factorize_fates(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{FATE_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{FATE_CODE} factorized according to the default sequence of fate codes
}
\description{
Factorizes the fate codes in a data frame
}
