% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fisheries}
\alias{factorize_fisheries}
\title{Factorizes the fishery codes in a data frame}
\usage{
factorize_fisheries(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{FISHERY_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{FISHERY_CODE} factorized according to the default sequence of fishery codes
}
\description{
Factorizes the fishery codes in a data frame
}
