% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fishery_groups}
\alias{factorize_fishery_groups}
\title{Factorizes the fishery group codes in a data frame}
\usage{
factorize_fishery_groups(to_factorize, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{to_factorize}{The data frame whose \code{FISHERY_GROUP_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{FISHERY_GROUP_CODE} factorized according to the default sequence of fishery group codes
}
\description{
Factorizes the fishery group codes in a data frame
}
