% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fishing_grounds}
\alias{factorize_fishing_grounds}
\title{Factorizes the fishing ground codes (\code{IRWESIO} and \code{IREASIO}) in a data frame}
\usage{
factorize_fishing_grounds(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{FISHING_GROUND_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{FISHING_GROUND_CODE} factorized according to the default sequence of fishing ground codes
}
\description{
Factorizes the fishing ground codes (\code{IRWESIO} and \code{IREASIO}) in a data frame
}
