% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_fleets}
\alias{factorize_fleets}
\title{Factorizes the fleet codes in a data frame}
\usage{
factorize_fleets(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{FLEET_CODE} column should be factorized}

\item{connection}{A connection to the IOTDB (required to retrieve fleet data)}
}
\value{
the original data frame with the \code{FLEET_CODE} factorized according to the default sequence of fleet codes
}
\description{
Factorizes the fleet codes in a data frame
}
