% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_gears}
\alias{factorize_gears}
\title{Factorizes the gear codes in a data frame}
\usage{
factorize_gears(to_factorize, connection = DB_IOTDB())
}
\arguments{
\item{to_factorize}{The data frame whose \code{GEAR_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{GEAR_CODE} factorized according to the default sequence of gear codes
}
\description{
Factorizes the gear codes in a data frame
}
