% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_qualities}
\alias{factorize_qualities}
\title{Factorizes the quality codes in a data frame.
This function is unnecessary, as the regular NC, EF, CA, CE and SA datasets do not come with a \code{QUALITY_CODE} column.
The only place where quality codes are leveraged is during the creation of dataset-specific quality charts, which include already
the logic to factorize the colors by quality code.}
\usage{
factorize_qualities(data)
}
\arguments{
\item{data}{The data frame whose \code{QUALITY_CODE} column should be factorized}

\item{connection}{A connection to the IOTDB (required for the factorization of fleets)}
}
\value{
the original data frame with the \code{QUALITY_CODE} factorized according to the default sequence of quality codes
}
\description{
Factorizes the quality codes in a data frame.
This function is unnecessary, as the regular NC, EF, CA, CE and SA datasets do not come with a \code{QUALITY_CODE} column.
The only place where quality codes are leveraged is during the creation of dataset-specific quality charts, which include already
the logic to factorize the colors by quality code.
}
