% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_quarters}
\alias{factorize_quarters}
\title{Factorizes the quarters in a data frame}
\usage{
factorize_quarters(data)
}
\arguments{
\item{data}{The data frame whose \code{QUARTER} column should be factorized}
}
\value{
the original data frame with the \code{QUARTER} factorized according to the default sequence of quarters
}
\description{
Factorizes the quarters in a data frame
}
