% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_ROS_event_types}
\alias{factorize_ROS_event_types}
\title{Factorizes the ROS event type codes in a data frame}
\usage{
factorize_ROS_event_types(data)
}
\arguments{
\item{data}{The data frame whose \code{EVENT_TYPE_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{EVENT_TYPE_CODE} factorized according to the default sequence of event type codes
}
\description{
Factorizes the ROS event type codes in a data frame
}
