% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_species}
\alias{factorize_species}
\title{Factorizes the species codes in a data frame}
\usage{
factorize_species(
  to_factorize,
  collapse_secondary_species = FALSE,
  connection = iotc.core.db.connections::getDefaultDBIHandler()()
)
}
\arguments{
\item{to_factorize}{The data frame whose \code{SPECIES_CODE} column should be factorized}

\item{collapse_secondary_species}{If \code{TRUE} all species other than the main ones will be assigned to \code{UNCL} / \code{Unclassified}}

\item{connection}{A connection to the \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{SPECIES_CODE} factorized according to the default sequence of species codes
}
\description{
Factorizes the species codes in a data frame
}
