% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_species_wps}
\alias{factorize_species_wps}
\title{Factorizes the species working party codes in a data frame}
\usage{
factorize_species_wps(data, connection = DB_IOTDB())
}
\arguments{
\item{data}{The data frame whose \code{SPECIES_WP_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{SPECIES_WP_CODE} factorized according to the default sequence of species group codes
}
\description{
Factorizes the species working party codes in a data frame
}
