% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FATE_CODE}
\alias{FATE_CODE}
\title{The fate code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FATE_CODE
}
\description{
The fate code column constant
}
\keyword{datasets}
