% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FATE_TYPE}
\alias{FATE_TYPE}
\title{The fate type column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FATE_TYPE
}
\description{
The fate type column constant
}
\keyword{datasets}
