% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FIRST_CLASS_LOW}
\alias{FIRST_CLASS_LOW}
\title{The 'first class low' column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FIRST_CLASS_LOW
}
\description{
The 'first class low' column constant
}
\keyword{datasets}
