% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISH_COUNT}
\alias{FISH_COUNT}
\title{The fish count column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISH_COUNT
}
\description{
The fish count column constant
}
\keyword{datasets}
