% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISHERY_CODE}
\alias{FISHERY_CODE}
\title{The fishery code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISHERY_CODE
}
\description{
The fishery code column constant
}
\keyword{datasets}
