% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISHERY_GROUP}
\alias{FISHERY_GROUP}
\title{The fishery group column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISHERY_GROUP
}
\description{
The fishery group column constant
}
\keyword{datasets}
