% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISHERY_TYPE_CODE}
\alias{FISHERY_TYPE_CODE}
\title{The fishery type code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISHERY_TYPE_CODE
}
\description{
The fishery type code column constant
}
\keyword{datasets}
