% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISHING_GROUND}
\alias{FISHING_GROUND}
\title{The fishing ground column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISHING_GROUND
}
\description{
The fishing ground column constant
}
\keyword{datasets}
