% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FISHING_GROUND_CODE}
\alias{FISHING_GROUND_CODE}
\title{The fishing ground code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FISHING_GROUND_CODE
}
\description{
The fishing ground code column constant
}
\keyword{datasets}
