% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{FLAG_CODE}
\alias{FLAG_CODE}
\title{The flag code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
FLAG_CODE
}
\description{
The flag code column constant
}
\keyword{datasets}
