% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fleet_codes.R
\name{fleet_name_}
\alias{fleet_name_}
\title{DEPRECATED
Extracts a fleet name from the \code{CountryStratVsFleet} table on \code{IOTDB} for a given fleet code}
\usage{
fleet_name_(fleet_code, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{fleet_code}{A fleet code}

\item{connection}{A connection to the IOTDB}
}
\value{
The English name for the provided fleet code
}
\description{
DEPRECATED
Extracts a fleet name from the \code{CountryStratVsFleet} table on \code{IOTDB} for a given fleet code
}
\examples{
fleet_name("EU_FRA")
}
