% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{GEAR_CODE}
\alias{GEAR_CODE}
\title{The gear code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
GEAR_CODE
}
\description{
The gear code column constant
}
\keyword{datasets}
