% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{IUCN_STATUS}
\alias{IUCN_STATUS}
\title{The 'IUCN status' column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
IUCN_STATUS
}
\description{
The 'IUCN status' column constant
}
\keyword{datasets}
