% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{LENGTH_MEASURE_UNIT_CODE}
\alias{LENGTH_MEASURE_UNIT_CODE}
\title{The length measure unit code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
LENGTH_MEASURE_UNIT_CODE
}
\description{
The length measure unit code column constant
}
\keyword{datasets}
