% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{MEASURE_TYPE}
\alias{MEASURE_TYPE}
\title{The measure type codecolumn constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
MEASURE_TYPE
}
\description{
The measure type codecolumn constant
}
\keyword{datasets}
