% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{MONTH_END}
\alias{MONTH_END}
\title{The month (end) column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
MONTH_END
}
\description{
The month (end) column constant
}
\keyword{datasets}
