% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{NC.raw}
\alias{NC.raw}
\title{Alias for \code{\link{NC_raw}}}
\usage{
NC.raw(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{NC_raw}}
}
