% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{NC_raw}
\alias{NC_raw}
\title{Retrieves *raw* yearly nominal catch data from the 'standard' \code{V_LEGACY_NC_STD} or 'official' \code{V_LEGACY_NC_OFF} view on IOTDB, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria.}
\usage{
NC_raw(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\arguments{
\item{use_official}{A flag that tells whether using the official (if TRUE) or standard (if FALSE) view}

\item{years}{A vector of years the extraction should be limited to}

\item{flag_codes}{A vector of flag codes the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to#'}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{fishery_group_codes}{A vector of fishery group codes the extraction should be limited to}

\item{species_codes}{A vector of species codes the extraction should be limited to}

\item{species_category_codes}{A vector of species category codes (BILLFISH, SEERFISH, NERITIC, TEMPERATE, TROPICAL, TUNAS_NEI, SHARKS, MANTAS, SEABIRDS, CETACEANS, TURTLES, OTHERS) the extraction should be limited to}

\item{species_group_codes}{A vector of species group codes (BILLFISH, SEERFISH, TUNAS, SHARKS, OTHERS) the extraction should be limited to}

\item{species_wp_codes}{A vector of species WP codes (BILL, NERI, TEMP, TROP, BYCT, UNCL) the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{IOTDB}} database}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}
}
\value{
a data frame with the records from \code{V_LEGACY_NC} matching the filtering criteria
}
\description{
The 'standard' version includes the 'FLAG_CODE' column as a combination of the codes from both
the flag country and the reporting country for the record, and the 'FLEET_CODE' column as it derives
from the combination of FLAG_CODE and fishing gear mapped by the \code{CountryStratVsFleet}
(which turns several flag + reporting country into generic fleets such as \code{NEICE}, \code{NEIFR}, \code{NEIPS} and similar).
The 'official' version only includes the 'FLEET_CODE' column determined as indicated above.
}
\examples{
NC_raw(years = seq(1980, 2000, 1), species_codes = c("ALB", "SBF"))
NC_raw(use_official = FALSE, years = seq(1980, 2000, 1), species_codes = c("ALB", "SBF"))
}
