% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{NC_raw_std}
\alias{NC_raw_std}
\title{Same as \code{NC_raw(use_official = FALSE, ...)}}
\usage{
NC_raw_std(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  connection = DB_IOTDB(),
  factorize_result = TRUE
)
}
\description{
Same as \code{NC_raw(use_official = FALSE, ...)}
}
