% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_buoys_import.R
\name{persist_buoy_data}
\alias{persist_buoy_data}
\title{Persists a dataset in the target database (generally, \code{\link{BUOY_DATA}})}
\usage{
persist_buoy_data(data, delete_existing = FALSE, connection = DB_BUOYS())
}
\arguments{
\item{delete_existing}{To remove records for the same strata from the target table (\code{BUOY_DATA}) before importing the data}

\item{connection}{A connection to the \code{\link{BUOY_DATA}} database}
}
\description{
Persists a dataset in the target database (generally, \code{\link{BUOY_DATA}})
}
