% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_buoys_import.R
\name{process_buoy_data_in_file}
\alias{process_buoy_data_in_file}
\title{Processes all buoy data in a file  folder, storing the result in the \code{\link{BUOY_DATA}} database}
\usage{
process_buoy_data_in_file(index = NA, file_path)
}
\arguments{
\item{index}{An optional index (ordinal value) for the file being processed. Makes sense only when invoked by the \code{\link{process_buoy_data_in_folder}} function
and is used for logging reasons only}

\item{file_path}{The path of the source Excel file (standard IOTC format)}
}
\description{
Processes all buoy data in a file  folder, storing the result in the \code{\link{BUOY_DATA}} database
}
