% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_buoys_import.R
\name{process_buoy_data_in_folder}
\alias{process_buoy_data_in_folder}
\title{Processes all buoy data in a given folder, storing the result in the \code{\link{BUOY_DATA}} database}
\usage{
process_buoy_data_in_folder(folder_path)
}
\arguments{
\item{folder_path}{The path of the folder containing the source Excel file (standard IOTC format)}
}
\description{
Processes all buoy data in a given folder, storing the result in the \code{\link{BUOY_DATA}} database
}
