% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{QUALITY_CODE}
\alias{QUALITY_CODE}
\title{The 'Quality code' column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
QUALITY_CODE
}
\description{
The 'Quality code' column constant
}
\keyword{datasets}
