% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{QUARTER}
\alias{QUARTER}
\title{The quarter column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
QUARTER
}
\description{
The quarter column constant
}
\keyword{datasets}
