% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{RAISE_CODE}
\alias{RAISE_CODE}
\title{The raise code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
RAISE_CODE
}
\description{
The raise code column constant
}
\keyword{datasets}
