% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raised.R
\name{raised.NC}
\alias{raised.NC}
\title{Alias for \code{\link{NC_raised}}}
\usage{
raised.NC(
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  connection = DB_WP_CE_RAISED(),
  connection_IOTDB = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{NC_raised}}
}
