% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{raw.CA}
\alias{raw.CA}
\title{Alias for \code{\link{CA_raw}}}
\usage{
raw.CA(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  catch_unit_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{CA_raw}}
}
