% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{raw.NC}
\alias{raw.NC}
\title{Alias for \code{\link{NC_raw}}}
\usage{
raw.NC(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{NC_raw}}
}
