% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_helpers.R
\name{reset_codelists_cache}
\alias{reset_codelists_cache}
\title{Resets the content of the codelists cache, triggering a reload from the database
the first time a codelist is accessed}
\usage{
reset_codelists_cache()
}
\description{
Resets the content of the codelists cache, triggering a reload from the database
the first time a codelist is accessed
}
