% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_lw_equations.R
\name{reset_equations_cache}
\alias{reset_equations_cache}
\title{Resets the content of the codelists cache, triggering a reload from the database
the first time a codelist is accessed}
\usage{
reset_equations_cache()
}
\description{
Resets the content of the codelists cache, triggering a reload from the database
the first time a codelist is accessed
}
