% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_ROS.R
\name{ros.EF}
\alias{ros.EF}
\title{Alias for \code{\link{EF_ros}}}
\usage{
ros.EF(
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  effort_unit_codes = NULL,
  connection = DB_ROS(),
  IOTDB_connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{EF_ros}}
}
