% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_ROS.R
\name{ros.raw.SETS}
\alias{ros.raw.SETS}
\title{Alias for \code{\link{SETS_raw_ros}}}
\usage{
ros.raw.SETS(
  years = NULL,
  date_from = NULL,
  date_to = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  event_type_codes = NULL,
  connection = DB_ROS(),
  IOTDB_connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{SETS_raw_ros}}
}
