% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{SAMPLE_SIZE}
\alias{SAMPLE_SIZE}
\title{The sample size column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
SAMPLE_SIZE
}
\description{
The sample size column constant
}
\keyword{datasets}
