% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_ROS.R
\name{SETS_raw_ros}
\alias{SETS_raw_ros}
\title{Retrieves ROS raw set data (including the type of event among \code{SETTING} and \code{HAULING}, the starting time and coordinates, the ending time and coordinates and the effort measured for the event)}
\usage{
SETS_raw_ros(
  years = NULL,
  date_from = NULL,
  date_to = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  event_type_codes = NULL,
  connection = DB_ROS(),
  IOTDB_connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\arguments{
\item{years}{A vector of years the extraction should be limited to}

\item{date_from}{A starting date (as '\code{yyyy-MM-dd}') the extraction should be limited to}

\item{date_to}{An ending date (as '\code{yyyy-MM-dd}') the extraction should be limited to}

\item{flag_codes}{A vector of flag codes the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{fishery_group_codes}{A vector of fishery group codes the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{ROS}} database}

\item{IOTDB_connection}{An ODBC connection to the \code{\link{IOTDB}} database to support retrieval of reference data}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}

\item{school_type_codes}{A vector of school type codes the extraction should be limited to}

\item{species_codes}{A vector of species codes the extraction should be limited to}

\item{species_category_codes}{A vector of species category codes (BILLFISH, SEERFISH, NERITIC, TEMPERATE, TROPICAL, TUNAS_NEI, SHARKS, MANTAS, SEABIRDS, CETACEANS, TURTLES, OTHERS) the extraction should be limited to}

\item{species_group_codes}{A vector of species group codes (BILLFISH, SEERFISH, TUNAS, SHARKS, OTHERS) the extraction should be limited to}

\item{species_wp_codes}{A vector of species WP codes (BILL, NERI, TEMP, TROP, BYCT, UNCL) the extraction should be limited to}

\item{length_measure_type_codes}{A vector of length measure type codes the extraction should be limited to}

\item{weight_measure_type_codes}{A vector of length measure type codes the extraction should be limited to}

\item{sex_codes}{A vector of sex codes the extraction should be limited to}
}
\value{
a data frame with the records from \code{V_SETS_RAW} matching the filtering criteria
}
\description{
Retrieves ROS raw set data (including the type of event among \code{SETTING} and \code{HAULING}, the starting time and coordinates, the ending time and coordinates and the effort measured for the event)
}
