% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{SEX_CODE}
\alias{SEX_CODE}
\title{The sex code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
SEX_CODE
}
\description{
The sex code column constant
}
\keyword{datasets}
