% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw.R
\name{SF.raw}
\alias{SF.raw}
\title{Alias for \code{\link{SF_raw}}}
\usage{
SF.raw(
  use_official = TRUE,
  years = NULL,
  flag_codes = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  measure_type_codes = NULL,
  sex_codes = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_results = TRUE
)
}
\description{
Alias for \code{\link{SF_raw}}
}
