% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_est.R
\name{SF_est}
\alias{SF_est}
\title{Retrieves *estimated* size-frequency data from the \code{vwSF<species_code>} view on IOTDB, optionally filtering the records
according to a variable set of user-provided, AND-ed criteria}
\usage{
SF_est(
  species_code,
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  connection = DB_IOTDB(),
  factorize_results = TRUE
)
}
\arguments{
\item{species_code}{A species code the extraction should be limited to.}

\item{years}{A vector of years the extraction should be limited to}

\item{fleet_codes}{A vector of flag codes the extraction should be limited to}

\item{gear_codes}{A vector of gear codes the extraction should be limited to#'}

\item{fishery_codes}{A vector of fishery codes the extraction should be limited to}

\item{connection}{An ODBC connection to the \code{\link{IOTDB}} database}

\item{factorize_results}{Whether the result data table has to be factorized. Among other things, when set to \code{TRUE} maps all less-important species to \code{UNCL} (within their species group)}

\item{fishery_groups}{A vector of fishery groups the extraction should be limited to
It can be one among: \code{ALB}, \code{BET}, \code{SKJ}, \code{YFT}, \code{BLM}, \code{BUM}, \code{MLS}, \code{SFA}, \code{SWO},
\code{BLT}, \code{COM}, \code{FRI}, \code{GUT}, \code{KAW}, \code{LOT} or \code{SKH}}
}
\value{
a data frame with the records from \code{vwSF<species_code>} matching the filtering criteria
}
\description{
It only includes the 'FLEET_CODE' column as it derives from the combination of FLAG_CODE (flag country + reporting country and
fishing gear mapped by the \code{CountryStratVsFleet} (which turns several flag + reporting country into generic fleets such as \code{NEICE}, \code{NEIFR}, \code{NEIPS} and similar).
The 'official' version only includes the 'FLEET_CODE' column determined as indicated above.
}
\examples{
SF_est(years = seq(1980, 2000, 1), species_code = "ALB")
}
