% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_datasets_raw_helpers.R
\name{SF_raw_off}
\alias{SF_raw_off}
\title{Same as \code{SF_raw(use_official = TRUE, ...)}}
\usage{
SF_raw_off(
  years = NULL,
  fleet_codes = NULL,
  gear_codes = NULL,
  fishery_codes = NULL,
  fishery_group_codes = NULL,
  school_type_codes = NULL,
  species_codes = NULL,
  species_category_codes = NULL,
  species_group_codes = NULL,
  species_wp_codes = NULL,
  measure_type_codes = NULL,
  sex_codes = NULL,
  connection = iotc.core.db.connections::getDefaultDBIHandler()(),
  factorize_result = TRUE
)
}
\description{
Same as \code{SF_raw(use_official = TRUE, ...)}
}
