% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{SIZE_INTERVAL}
\alias{SIZE_INTERVAL}
\title{The size class column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
SIZE_INTERVAL
}
\description{
The size class column constant
}
\keyword{datasets}
