% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{SPECIES_GROUP_CODE}
\alias{SPECIES_GROUP_CODE}
\title{The species group code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
SPECIES_GROUP_CODE
}
\description{
The species group code column constant
}
\keyword{datasets}
