% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_species_codes.R
\name{species_name}
\alias{species_name}
\title{Extracts a species name from the \code{cdeSpecies} table on \code{IOTDB} for a given species code}
\usage{
species_name(species_code, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{species_code}{A species code}

\item{connection}{A connection to the IOTDB}
}
\value{
The English name for the provided species code
}
\description{
Extracts a species name from the \code{cdeSpecies} table on \code{IOTDB} for a given species code
}
\examples{
species_name("YFT")
}
