% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{SPECIES_WP}
\alias{SPECIES_WP}
\title{The working party column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
SPECIES_WP
}
\description{
The working party column constant
}
\keyword{datasets}
