% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{TOT_KG_FISH}
\alias{TOT_KG_FISH}
\title{The tot kg fish column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
TOT_KG_FISH
}
\description{
The tot kg fish column constant
}
\keyword{datasets}
