% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fishery_codes.R
\name{update_fishery_codes_for_species}
\alias{update_fishery_codes_for_species}
\title{Adds fishery codes for a specific species to provided data frame according to the \code{FISHERY_GROUP} column}
\usage{
update_fishery_codes_for_species(data)
}
\description{
Adds fishery codes for a specific species to provided data frame according to the \code{FISHERY_GROUP} column
}
