% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fishery_codes.R
\name{update_fishery_group_codes_for_species}
\alias{update_fishery_group_codes_for_species}
\title{Marks species-specific minor fisheries groups as 'OT'}
\usage{
update_fishery_group_codes_for_species(data)
}
\description{
Marks species-specific minor fisheries groups as 'OT'
}
