% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_mgmt_fishery_codes.R
\name{update_fishery_group_codes_for_species_group}
\alias{update_fishery_group_codes_for_species_group}
\title{Marks species group-specific minor fisheries groups as 'OT'}
\usage{
update_fishery_group_codes_for_species_group(
  data,
  species_group = WP_TROP,
  connection = DB_IOTDB()
)
}
\description{
Marks species group-specific minor fisheries groups as 'OT'
}
