% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{WEIGHT_MEASURE_UNIT_CODE}
\alias{WEIGHT_MEASURE_UNIT_CODE}
\title{The weight measure unit code column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
WEIGHT_MEASURE_UNIT_CODE
}
\description{
The weight measure unit code column constant
}
\keyword{datasets}
