% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_constants.R
\docType{data}
\name{YEAR}
\alias{YEAR}
\title{The year column constant}
\format{
An object of class \code{character} of length 1.
}
\usage{
YEAR
}
\description{
The year column constant
}
\keyword{datasets}
