% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_data_factors.R
\name{factorize_raisings}
\alias{factorize_raisings}
\title{Factorizes the raising codes in a data frame}
\usage{
factorize_raisings(to_factorize, connection = iotc.core.db.connections::getDefaultDBIHandler()())
}
\arguments{
\item{to_factorize}{The data frame whose \code{RAISE_CODE} column should be factorized}

\item{connection}{A connection to \code{\link{IOTDB}}}
}
\value{
the original data frame with the \code{RAISE_CODE} factorized according to the default sequence of raising codes
}
\description{
Factorizes the raising codes in a data frame
}
